src/main/java/domain/menu/Menu.java
package domain.menu;

import domain.enums.Category;
import domain.vo.Quantity;

import java.util.Objects;

public class Menu {
    private final int number;
    private final String name;
    private final Category category;
    private final int price;

    public Menu(final int number, final String name, final Category category, final int price) {
        this.number = number;
        this.name = name;
        this.category = category;
        this.price = price;
    }

    public String getName() {
        return name;
    }

    public int getPrice() {
        return price;
    }

    public int getMultiplyPriceOf(final Quantity quantity) {
        return this.price * quantity.get();
    }

    public boolean equalsOf(Category category) {
        return this.category.equals(category);
    }

    public boolean equalsOf(final int number) {
        return this.number == number;
    }

    @Override
    public boolean equals(final Object o) {
        if (this == o) return true;
        if (!(o instanceof Menu)) return false;
        final Menu menu = (Menu) o;
        return number == menu.number;
    }

    @Override
    public int hashCode() {
        return Objects.hash(number);
    }

    @Override
    public String toString() {
        return category + " " + number + " - " + name + " : " + price + "원";
    }
}
